refactor: split frontend api and pipeline action helpers
